add ControlledVocabularyInput component to GenericWork schema
