+++ package.json
@@ -45,7 +45,7 @@
     "rc-slider": "^5.1.2",
     "react": "^15.3.1",
     "react-addons-shallow-compare": "^15.3.1",
-    "react-autocomplete": "^1.2.1",
+    "react-autocomplete": "^1.4.0",
     "react-dom": "^15.1.0",
     "react-dropzone": "^3.5.3",
     "react-modal": "^1.4.0",
@@ -88,7 +88,7 @@
     "react-addons-test-utils": "^15.3.1",
     "redux-mock-store": "^1.2.0"
   },
-	"plugins": [
-		"karma-mocha"
-	]
+  "plugins": [
+    "karma-mocha"
+  ]
 }
